room entitites added!

+++ app/src/main/java/com/jnfran92/kotcoin/presentation/crypto/dataflow/action/CryptoDetailsAction.kt
@@ -1,5 +1,5 @@
 package com.jnfran92.kotcoin.presentation.crypto.dataflow.action
 
 sealed class CryptoDetailsAction {
-    class GetCryptoDetails(val itemId: Int) : CryptoDetailsAction()
+    class GetCryptoDetails(val itemId: Long) : CryptoDetailsAction()
 }
