+++ app/src/main/java/com/jnfran92/kotcoin/presentation/crypto/dataflow/processor/CryptoDetailsProcessor.kt
@@ -28,7 +28,7 @@ class CryptoDetailsProcessor @Inject constructor(
     }
 
 
-    private fun buildGetCryptoDetailsUseCase(itemId: Int): Observable<CryptoDetailsResult>? {
+    private fun buildGetCryptoDetailsUseCase(itemId: Long): Observable<CryptoDetailsResult>? {
         return getCryptoDetailsUseCase
             .toSingle(itemId)
             .map(toUIMapper::transform)
